Separated CourseDetailsScreen page into separate folder
