+++ BuckyClass/src/screens/CourseDetail/CourseDetailsScreen_CSS.ts
@@ -1,0 +1,301 @@
+import { StyleSheet, Dimensions } from "react-native";
+
+const screenWidth = Dimensions.get("window").width;
+
+const styles = StyleSheet.create({
+    safeArea: {
+        flex: 1,
+    },
+    gradientBackground: {
+        flex: 1,
+        backgroundColor: "rgba(232, 221, 253, 0.60)",
+        borderWidth: 1,
+        borderColor: "rgba(255, 255, 255, 0.20)",
+        borderRadius: 30,
+    },
+    gradientStyle: {
+        flex: 1,
+        borderWidth: 1,
+        borderColor: "rgba(255, 255, 255, 0.20)",
+    },
+    blurOverlay: {
+        flex: 1,
+        backgroundColor: "rgba(255, 255, 255, 0.1)",
+        shadowColor: "rgba(0, 0, 0, 0.25)",
+        shadowOffset: { width: 0, height: 4 },
+        shadowOpacity: 1,
+        shadowRadius: 4,
+        elevation: 8,
+        paddingTop: 20,
+        paddingHorizontal: 20,
+        paddingBottom: 0,
+    },
+    scrollContent: {
+        padding: 16,
+    },
+    centeredContainer: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+        padding: 20,
+    },
+    loadingText: {
+        marginTop: 10,
+        fontFamily: "Nunito",
+        color: "#555",
+    },
+    courseName: {
+        fontSize: 20,
+        fontWeight: "bold",
+        marginBottom: 8,
+        fontFamily: "Nunito-ExtraBold",
+        color: "#171717",
+    },
+    courseInfoRow: {
+        flexDirection: "row",
+        justifyContent: "space-between",
+        marginBottom: 24,
+    },
+    courseInfoText: {
+        fontSize: 14,
+        color: "#777",
+        fontFamily: "Nunito",
+    },
+    sectionContainer: {
+        backgroundColor: "#fff",
+        padding: 16,
+        borderRadius: 24,
+        marginBottom: 12,
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOffset: { width: 0, height: 2 },
+        shadowRadius: 6,
+        shadowOpacity: 0.3,
+        elevation: 3,
+    },
+    sectionTitle: {
+        fontSize: 18,
+        fontWeight: "bold",
+        marginBottom: 16,
+        fontFamily: "Nunito-ExtraBold",
+        color: "#171717",
+    },
+    totalStudents: {
+        fontSize: 14,
+        marginBottom: 16,
+    },
+    chartContainer: {
+        alignItems: "center",
+        marginBottom: 16,
+    },
+    gradeDetailsContainer: {
+        flexDirection: "row",
+        flexWrap: "wrap",
+        justifyContent: "space-between",
+    },
+    gradeDetail: {
+        width: "48%",
+        marginBottom: 8,
+        fontSize: 14,
+    },
+    reviewItem: {
+        padding: 12,
+        borderBottomWidth: 1,
+        borderBottomColor: "#eee",
+        marginBottom: 12,
+    },
+    reviewHeader: {
+        flexDirection: "row",
+        justifyContent: "space-between",
+        alignItems: "center",
+        marginBottom: 8,
+    },
+    ratingContainer: {
+        flexDirection: "row",
+        alignItems: "center",
+    },
+    starIcon: {
+        marginRight: 2,
+    },
+    ratingText: {
+        marginLeft: 4,
+        fontSize: 14,
+        fontWeight: "bold",
+        color: "#333",
+    },
+    reviewDate: {
+        fontSize: 12,
+        color: "#999",
+        fontStyle: "italic",
+    },
+    reviewComment: {
+        fontSize: 14,
+        color: "#333",
+        lineHeight: 20,
+        marginBottom: 8,
+    },
+    reviewFooter: {
+        flexDirection: "row",
+        justifyContent: "flex-end",
+    },
+    likeContainer: {
+        flexDirection: "row",
+        alignItems: "center",
+    },
+    likeCount: {
+        marginLeft: 4,
+        fontSize: 12,
+        color: "#666",
+    },
+    emptyText: {
+        fontStyle: "italic",
+        color: "#999",
+    },
+    errorText: {
+        color: "#F97CBD",
+        fontSize: 16,
+        fontFamily: "Nunito-Bold",
+        textAlign: "center",
+        marginBottom: 20,
+    },
+    loginButton: {
+        backgroundColor: "#F97CBD",
+        borderRadius: 24,
+        padding: 12,
+        marginTop: 20,
+        alignItems: "center",
+        alignSelf: "center",
+        paddingHorizontal: 40,
+    },
+    loginButtonText: {
+        color: "#FFF",
+        fontSize: 16,
+        fontFamily: "Nunito-Bold",
+    },
+    averageGpa: {
+        fontSize: 16,
+        fontWeight: "bold",
+        color: "#4CAF50",
+        marginTop: 16,
+        fontFamily: "Nunito-Bold",
+    },
+    instructorName: {
+        fontSize: 14,
+        marginBottom: 8,
+        color: "#555",
+        fontFamily: "Nunito",
+    },
+    reviewUsername: {
+        fontSize: 14,
+        fontWeight: "bold",
+        color: "#555",
+        marginBottom: 4,
+        fontFamily: "Nunito-Bold",
+    },
+    reviewsContainer: {
+        marginTop: 8,
+    },
+    reviewCard: {
+        backgroundColor: "#f8f9fa",
+        borderRadius: 16,
+        padding: 16,
+        marginBottom: 12,
+        borderWidth: 1,
+        borderColor: "#eee",
+    },
+    reviewCardHeader: {
+        flexDirection: "row",
+        justifyContent: "space-between",
+        alignItems: "flex-start",
+        marginBottom: 12,
+    },
+    reviewUser: {
+        flexDirection: "row",
+        alignItems: "center",
+    },
+    reviewUserAvatar: {
+        width: 36,
+        height: 36,
+        borderRadius: 18,
+        backgroundColor: "#F97CBD",
+        justifyContent: "center",
+        alignItems: "center",
+        marginRight: 10,
+    },
+    reviewUserAvatarText: {
+        color: "white",
+        fontSize: 16,
+        fontWeight: "bold",
+        fontFamily: "Nunito-Bold",
+    },
+    reviewComment: {
+        fontSize: 14,
+        color: "#333",
+        lineHeight: 20,
+        marginBottom: 12,
+        fontFamily: "Nunito",
+    },
+    reviewAction: {
+        flexDirection: "row",
+        alignItems: "center",
+        marginRight: 16,
+    },
+    reviewActionText: {
+        marginLeft: 5,
+        fontSize: 12,
+        color: "#777",
+        fontFamily: "Nunito",
+    },
+    emptyReviewsContainer: {
+        alignItems: "center",
+        justifyContent: "center",
+        padding: 30,
+    },
+    emptySubText: {
+        fontSize: 14,
+        color: "#aaa",
+        marginTop: 5,
+        fontFamily: "Nunito",
+    },
+    buttonsContainer: {
+        flexDirection: "row",
+        justifyContent: "space-between",
+        marginBottom: 24,
+    },
+    actionButton: {
+        flexDirection: "row",
+        backgroundColor: "#fff",
+        borderRadius: 24,
+        padding: 12,
+        alignItems: "center",
+        justifyContent: "center",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOffset: { width: 0, height: 2 },
+        shadowRadius: 6,
+        shadowOpacity: 0.3,
+        elevation: 3,
+        borderWidth: 1,
+        borderColor: "#8863e4",
+        flex: 1,
+        marginHorizontal: 6,
+    },
+    actionButtonText: {
+        color: "#8863e4",
+        fontSize: 15,
+        fontWeight: "500",
+        fontFamily: "Nunito-Bold",
+        marginLeft: 6,
+    },
+    reviewButton: {
+        backgroundColor: "#F97CBD",
+        borderColor: "#F97CBD",
+    },
+    reviewButtonText: {
+        color: "#fff",
+        fontSize: 15,
+        fontWeight: "500",
+        fontFamily: "Nunito-Bold",
+        marginLeft: 6,
+    },
+});
+
+export default styles;
